+++ src/app/Recipe_List/RecipeList.test.tsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom';
 import { act, render, screen } from '@testing-library/react';
-import RecipeList from './RecipeList';
+import RecipeList from './page';
 
 describe('Recipe List', () => {
   const mockRecipes = {
